Move non expect/actual extension functions to respective *Ext files
